Fixed account not loading fast enough
Done by using a callback. The code is quite ugly though, we should consider
adding the score to the records in the database to simplify the queries
