+++ QRAdventure/app/src/main/java/com/example/qradventure/AccountCallback.java
@@ -1,0 +1,7 @@
+package com.example.qradventure;
+
+public interface AccountCallback {
+
+    public void toActivity(Boolean alreadyCreated);
+
+}
