+++ QRAdventure/app/src/main/java/com/example/qradventure/QueryHandler.java
@@ -3,11 +3,13 @@ package com.example.qradventure;
 import android.content.Intent;
 import android.provider.Settings;
 import android.util.Log;
+import android.widget.Toast;
 
 import androidx.annotation.NonNull;
 
 import com.google.android.gms.tasks.OnCompleteListener;
 import com.google.android.gms.tasks.Task;
+import com.google.firebase.firestore.DocumentSnapshot;
 import com.google.firebase.firestore.FirebaseFirestore;
 import com.google.firebase.firestore.QueryDocumentSnapshot;
 import com.google.firebase.firestore.QuerySnapshot;
@@ -26,6 +28,105 @@ public class QueryHandler {
     public QueryHandler(){
         FirebaseFirestore.getInstance();
     }
+
+
+    public void getLoginAccount(String androidDeviceID, AccountCallback callback){
+
+        db = FirebaseFirestore.getInstance();
+
+        db.collection("AccountDB")
+                .whereEqualTo("device_id", androidDeviceID)
+                .get()
+                .addOnCompleteListener(new OnCompleteListener<QuerySnapshot>() {
+                    @Override
+                    public void onComplete(@NonNull Task<QuerySnapshot> task) {
+                        if (task.isSuccessful()){
+                            // query completed
+                            // TODO: reformat this "for". Surely there's a better way than a for loop for 1 document??
+                            if (task.getResult().size() == 0) {
+                                // no documents found! go to registration
+                                Log.d("logs", "doc dne!");
+                                // disable backward navigation to this activity
+                                callback.toActivity(false);
+                            }
+                            for(QueryDocumentSnapshot doc: task.getResult()) {
+                                if (doc.exists()) {
+                                    //Fetch the user account
+                                    String email = (String) doc.getData().get("E-mail");
+                                    String loginQR = (String) doc.getData().get("LoginQR");
+                                    String phoneNumber = (String) doc.getData().get("Phone Number");
+                                    String statusQR = (String) doc.getData().get("StatusQR");
+                                    String username = (String) doc.getId();
+                                    Log.d("logs", "doc exists + " + username);
+
+                                    Account fetchedAccount =
+                                            new Account(username, email, phoneNumber, loginQR, statusQR);
+                                    CurrentAccount.setAccount(fetchedAccount);
+
+                                    // Account reconstructed - need to reconstruct records
+                                    Account account = CurrentAccount.getAccount();
+                                    try {
+                                        Log.d("logs", account.getUsername());
+                                        db.collection("AccountDB").document(CurrentAccount.getAccount().getUsername()).collection("My QR Records")
+                                                .get()
+                                                .addOnCompleteListener(new OnCompleteListener<QuerySnapshot>() {
+                                                    @Override
+                                                    public void onComplete(@NonNull Task<QuerySnapshot> task) {
+                                                        if (task.isSuccessful()) {
+                                                            for (QueryDocumentSnapshot document : task.getResult()) {
+                                                                String qrHash = (String) document.getData().get("QR");
+                                                                db.collection("QRDB").document(qrHash)
+                                                                        .get()
+                                                                        .addOnCompleteListener(new OnCompleteListener<DocumentSnapshot>() {
+                                                                            @Override
+                                                                            public void onComplete(@NonNull Task<DocumentSnapshot> task) {
+                                                                                if (task.isSuccessful()) {
+                                                                                    DocumentSnapshot document = task.getResult();
+                                                                                    String qrScore = "" + document.getData().get("Score");
+                                                                                    int qrValue = Integer.parseInt(qrScore);
+                                                                                    QR qr = new QR(qrHash, qrValue, null, null);
+                                                                                    Log.d("logs", qrHash + " " + qrValue);
+                                                                                    Record newRecord = new Record(account, qr);
+                                                                                    account.addRecord(newRecord);
+                                                                                } else {
+                                                                                    // ERROR: Query failed!
+                                                                                    Log.d("logs", "Cached get failed: ", task.getException());
+                                                                                }
+
+                                                                                callback.toActivity(true);
+                                                                            }
+                                                                        });
+                                                            }
+
+                                                        } else {
+                                                            // ERROR: query unsuccessful
+                                                            Log.d("logs", "Error getting documents: ", task.getException());
+                                                        }
+                                                    }
+                                                });
+                                    }
+                                    catch (Exception e) {
+                                        Log.d("OOPS", "Query failed");
+                                    }
+                                } else {
+                                    // doc does not exist. I think this case is covered already, but just incase:
+                                    Log.d("logs", "Document does not exist!");
+                                }
+                            }
+                        } else {
+                            // ERROR: query failed
+                            Log.d("logs", "DeviceID query failed!", task.getException());
+                        }
+
+
+
+                    }
+                });
+
+    }
+
+
+
 
 
     public void getOthersScanned(QR qr, QueryCallback myCallback){
